# expence-tracker/my-app/specs/1-expense-tracker-mvp/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Expense Tracker API
  description: RESTful API for the Expense Tracker MVP
  version: 1.0.0
  contact:
    name: API Support
    email: support@expensetracker.example.com
servers:
  - url: https://api.expensetracker.example.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /transactions:
    get:
      tags:
        - Transactions
      summary: List user transactions
      description: Get paginated list of transactions with optional filters
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [EXPENSE, INCOME]
        - name: category
          in: query
          schema:
            type: string
        - name: paymentMethod
          in: query
          schema:
            type: string
            enum: [CASH, CARD, BANK_TRANSFER, OTHER]
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Transactions
      summary: Create transaction
      description: Create a new expense or income transaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionInput'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /transactions/{id}:
    get:
      tags:
        - Transactions
      summary: Get transaction by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Transactions
      summary: Update transaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionInput'
      responses:
        '200':
          description: Transaction updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Transactions
      summary: Delete transaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      responses:
        '204':
          description: Transaction deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Categories
...
  /import/csv:
    post:
      tags:
        - Reports
      summary: Import transactions from CSV
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Import summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                  failed:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        row:
                          type: integer
                        error:
                          type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # User
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TransactionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [EXPENSE, INCOME]
        amount:
          type: number
          format: decimal
        category:
          type: string
        date:
          type: string
          format: date
        description:
          type: string
          maxLength: 200
          nullable: true
        paymentMethod:
          type: string
          enum: [CASH, CARD, BANK_TRANSFER, OTHER]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TransactionInput:
      type: object
      required:
        - type
        - amount
        - category
        - date
        - paymentMethod
      properties:
        type:
          type: string
          enum: [EXPENSE, INCOME]
        amount:
          type: number
          format: decimal
          minimum: 0.01
          maximum: 99999999.99
        category:
          type: string
          minLength: 1
          maxLength: 100
        date:
          type: string
          format: date
        description:
          type: string
          maxLength: 200
        paymentMethod:
          type: string
          enum: [CASH, CARD, BANK_TRANSFER, OTHER]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
        totalPages:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...